Always send sip presence containing sip address with either sip: or sips: prefix
